+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add Drupal GitlabCI
 - add cpsell project words for Gitlab-CI
 - add a new layer of performance by allowing developers to cache OCR'ed files
+- add command entity_to_text:tika:warmup
 
 ## [1.0.0] - 2023-01-27
 ### Added